Separate the reference computation and reference publishing
Now the contact based reference finder computes the reference poses
(ee and base) but only publishes them as ros topics in the ros frames.
Now base_reference_position_publisher pushes the (transformed) poses
to the px4 topic. In the future the latter will take the reference
and use it to create intermediate poses to ensure a desired (constant)
velocity of the base.