Use nominal instead of true joint state for contact based planning
The contact based planner now no longer has a subscriber to the joint state
but instead uses a newly create service (offered by the encoder node) that
provides the joint state within the constructor of the contact_based_planner.
This value during startup is then considered the nominal value for the
rest of the existance of the contact based planner.